shim: Add the Security Version Support
Security Version is a mechanism that allows the Linux distribution to
warn the user when a kernel with known security flaws is going to be
loaded. It relies on the following fields in the PE Image:

UINT16 MajorOperatingSystemVersion
UINT16 MinorOperatingSystemVersion
UINT16 MajorImageVersion
UINT16 MinorImageVersion

The four fields are redefined as:

Signer Name:      MajorOperatingSystemVersion + MinorOperatingSystemVersion
Distro Version:   MajorImageVersion
Security Version: MinorImageVersion

A Linux kernel for a certain distro can be identified with the signer
and the distro version. For example, a SUSE Linux Enterprise 12 SP2
kernel with security version 0:

Signer Name:      'S''U''S''E'
Distro Version:   1202
Security Version: 0

A white list is maintained in SVList (BS+NV) and the content will be
mirrored to SVListRT (BS+RT). The list structure is like this:

[Signer 1][Distro][Security][Distro][Security]
[Signer 2][Distro][Security]
...

The security version in SVList can only be increased monotonically
unless the user physically presents and agrees to lower the value.

When grub2 or other bootloader call shim_verify() to verify a Linux
kernel, shim matches the security version and allows the kernel to
boot only if its security version is higher than or equal to the
existed one. Otherwise, a prompt will show to warn the user.

(If SVList doesn't exist, the list will be created automatically based
 on the first verified kernel.)

SVList will be updated under the following circumstances:

1. If the verified kernel has the higher security version, update the
   security version in SVList.

2. If the verified kernel has the lower security version, a warning
   prompt will show to ask the user if the security version in SVList
   should be lowered or not.

3. If the verified kernel has the signer or distro version not existed
   in SVList, a warning prompt will show to ask the user if the new
   signer or distro version should be added into SVList.

4. To massively update SVList, the user can create SVListNew (BS+RT+NV)
   and SVListSig (BS+RT+NV) to notify shim. SVListNew contains the list
   to be updated while SVListSig stores the signature of SVListNew. If
   SVListNew + SVListSig can be verified by the built-in keys or MOKs,
   shim will merge SVListNew into SVList. Note: when merging SVListNew,
   only the higher security version or the new signer/distro version
   will be merged.

To make reinstallation easier, shim will reset SVList when it's loaded
from a USB stick, DVD, PXE, or HTTP Boot and then create a new SVList
with the newly installed kernel.

Security Version is disabled by default. To enable this function:

 $ make ENABLE_SV_VERIFY=1

Signed-off-by: Gary Lin <glin@suse.com>